# openapi.yaml
openapi: 3.1.0
info:
  title: Pipeline AI DevOps Platform API
  version: 1.0.0
  description: |
    AI DevOps automation platform for deploying applications to Vercel and Render.
    Provides sandboxed workspace execution via Daytona, AI-powered repository analysis,
    and automated deployment orchestration.
  contact:
    name: Pipeline Support
    email: support@pipeline.dev
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.pipeline.dev/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /repos/{repo_id}:
    get:
      summary: Get repository details
      tags: [Repositories]
      operationId: getRepo
      parameters:
        - name: repo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Repository details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepoConnection'
        '404':
          description: Repository not found

  # ==================== DEPLOYMENTS ====================
  /deployments:
    get:
      summary: List all deployments
      tags: [Deployments]
      operationId: listDeployments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, succeeded, failed, cancelled]
        - name: repo_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentList'

    post:
      summary: Create deployment plan
      description: Create a new deployment plan based on repository analysis
      tags: [Deployments]
      operationId: createDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentCreateRequest'
      responses:
        '201':
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentPlan'
        '400':
          description: Invalid request
  /deployments/{deployment_id}:
    get:
      summary: Get deployment status
      tags: [Deployments]
      operationId: getDeployment
      parameters:
        - name: deployment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentStatus'
        '404':
          description: Deployment not found
  /deployments/{deployment_id}/run:
    post:
      summary: Execute deployment
      description: |
        Execute the deployment in a Daytona sandbox. This will:
        1. Start the sandbox
        2. Run build commands
        3. Deploy services to Vercel/Render
        4. Track deployment status
      tags: [Deployments]
      operationId: runDeployment
      parameters:
        - name: deployment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Deployment execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentExecutionResponse'
        '400':
          description: Deployment not in runnable state
  /deployments/{deployment_id}/cancel:
    post:
      summary: Cancel deployment
      tags: [Deployments]
      operationId: cancelDeployment
      parameters:
        - name: deployment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deployment cancelled
  /deployments/{deployment_id}/logs:
    get:
      summary: Get deployment logs
      description: Retrieve logs from the Daytona sandbox for this deployment
      tags: [Deployments]
      operationId: getDeploymentLogs
      parameters:
        - name: deployment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tail
          in: query
          description: Number of log lines to return from the end
          schema:
            type: integer
            default: 100
        - name: since
          in: query
          description: Return logs since this timestamp (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Deployment logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentLogs'
  /deployments/{deployment_id}/retry:
    post:
      summary: Retry failed deployment
      tags: [Deployments]
      operationId: retryDeployment
      parameters:
        - name: deployment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Deployment retry initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentExecutionResponse'

  # ==================== SANDBOXES ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RepoConnection:
      type: object
      required: [id, repo_url, provider, status, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the connected repository
        repo_url:
          type: string
          format: uri
        provider:
          type: string
          enum: [github, gitlab]
        branch:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [pending, connected, analyzing, error]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    DeploymentCreateRequest:
      type: object
      required: [repo_id, services]
      properties:
        repo_id:
          type: string
          format: uuid
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceDeploymentConfig'
        environment:
          type: string
          enum: [development, staging, production]
          default: production
        branch:
          type: string
          description: Branch to deploy from
        env_variables:
          type: object
          additionalProperties:
            type: string
          description: Environment variables for all services
    ServiceDeploymentConfig:
      type: object
      required: [name, path, platform]
      properties:
        name:
          type: string
        path:
          type: string
        platform:
          type: string
          enum: [vercel, render, docker]
        build_command:
          type: string
        start_command:
          type: string
        output_directory:
          type: string
        env_variables:
          type: object
          additionalProperties:
            type: string
        vercel_config:
          $ref: '#/components/schemas/VercelDeploymentConfig'
        render_config:
          $ref: '#/components/schemas/RenderDeploymentConfig'
    VercelDeploymentConfig:
      type: object
      properties:
        project_name:
          type: string
          description: Vercel project name
        team_id:
          type: string
        framework:
          type: string
    RenderDeploymentConfig:
      type: object
      properties:
        service_name:
          type: string
        service_type:
          type: string
          enum: [web_service, static_site, background_worker, cron_job]
        plan:
          type: string
          default: starter
    DeploymentPlan:
      type: object
      required: [id, repo_id, status, services, created_at]
      properties:
        id:
          type: string
          format: uuid
        repo_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, planned, approved]
        services:
          type: array
          items:
            $ref: '#/components/schemas/PlannedService'
        environment:
          type: string
        branch:
          type: string
        created_at:
          type: string
          format: date-time
    PlannedService:
      type: object
      required: [name, path, platform]
      properties:
        name:
          type: string
        path:
          type: string
        platform:
          type: string
          enum: [vercel, render, docker]
        status:
          type: string
          enum: [pending, deploying, deployed, failed, skipped]
        estimated_duration_seconds:
          type: integer
          description: Estimated deployment duration
    DeploymentStatus:
      type: object
      required: [id, repo_id, status, services, created_at]
      properties:
        id:
          type: string
          format: uuid
        repo_id:
          type: string
          format: uuid
        sandbox_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled, retrying]
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceDeployment'
        environment:
          type: string
        branch:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_seconds:
          type: integer
        error_message:
          type: string
        retry_count:
          type: integer
          default: 0
    ServiceDeployment:
      type: object
      required: [name, path, platform, status]
      properties:
        name:
          type: string
        path:
          type: string
        platform:
          type: string
          enum: [vercel, render, docker]
        status:
          type: string
          enum: [pending, building, deploying, deployed, failed, skipped]
        build_logs_url:
          type: string
          format: uri
        deployment_url:
          type: string
          format: uri
        platform_deployment_id:
          type: string
          description: Deployment ID from the platform (Vercel/Render)
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        error_message:
          type: string
    DeploymentExecutionResponse:
      type: object
      required: [deployment_id, status, message]
      properties:
        deployment_id:
          type: string
          format: uuid
        status:
          type: string
        message:
          type: string
        estimated_duration_seconds:
          type: integer
    DeploymentLogs:
      type: object
      required: [deployment_id, logs, timestamp]
      properties:
        deployment_id:
          type: string
          format: uuid
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        timestamp:
          type: string
          format: date-time
    DeploymentList:
      type: object
      required: [deployments, total, limit, offset]
      properties:
        deployments:
          type: array
          items:
            $ref: '#/components/schemas/DeploymentStatus'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    LogEntry:
      type: object
      required: [timestamp, level, message]
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [debug, info, warning, error, fatal]
        message:
          type: string
        source:
          type: string
          description: Component that generated the log

    # ==================== SANDBOX SCHEMAS ====================
...